+++ packages/components/scripts/generateDSRComponents.ts
@@ -561,6 +561,11 @@ $&`
           ) // rewire source for linkEl
           .replace(/(href: linkEl\.href),/, '$1 || linkEl.to,') // fallback for framework links
           .replace(/{this\.props\.children}/, '{manipulatedChildren}'); // apply manipulated children
+      } else if (tagName === 'p-link-tile-product') {
+        newFileContent = newFileContent
+          .replace(/LinkTileProductAspectRatio,/, '')
+          .replace(/LinkTileProductLikeEvent,/, '')
+          .replace(/LinkTileProductTarget,/, '');
       }
 
       return newFileContent;