+++ packages/components/src/components/link-tile-product/link-tile-product-utils.ts
@@ -1,0 +1,13 @@
+import { type LinkTarget } from '../../utils/link-button/link-target';
+import { type TileAspectRatio } from '../../utils';
+
+export type LinkTileProductTarget = LinkTarget;
+export type LinkTileProductLikeEvent = {
+  liked: boolean;
+};
+
+export const TILE_PRODUCT_ASPECT_RATIOS = ['3:4', '9:16'] as const;
+export type LinkTileProductAspectRatio = Extract<TileAspectRatio, (typeof TILE_PRODUCT_ASPECT_RATIOS)[number]>;
+
+export const headerSlot = 'header';
+export const anchorSlot = 'anchor';