+++ packages/storefront/src/components/Markdown.vue
@@ -124,7 +124,7 @@
           &[id] {
             p-link-pure {
               visibility: hidden;
-              margin-left: $pds-spacing-static-small;
+              margin-inline-start: $pds-spacing-static-small;
             }
 
             &:hover {
@@ -194,8 +194,8 @@
 
         // Blockquote
         blockquote {
-          padding-left: $pds-spacing-static-medium;
-          border-left: 5px solid var(--theme-contrast-low);
+          padding-inline-start: $pds-spacing-static-medium;
+          border-inline-start: 5px solid var(--theme-contrast-low);
         }
 
         // Lists
@@ -203,7 +203,7 @@
         ol {
           @include pds-text-small;
           margin-top: $pds-spacing-fluid-medium;
-          padding-left: $pds-spacing-static-large;
+          padding-inline-start: $pds-spacing-static-large;
 
           ul,
           ol {
@@ -279,13 +279,13 @@
           }
 
           th {
-            text-align: left;
+            text-align: start;
             padding-bottom: $pds-spacing-static-small;
             border-bottom: 1px solid var(--theme-contrast-low);
           }
 
           td {
-            text-align: left;
+            text-align: start;
             padding: {
               top: $pds-spacing-static-small;
               bottom: $pds-spacing-static-small;
@@ -297,7 +297,7 @@
 
           th ~ th,
           td ~ td {
-            padding-left: $pds-spacing-static-medium;
+            padding-inline-start: $pds-spacing-static-medium;
           }
         }